WIP - introduce discount strategy introduce TwoForAmountOffer
