+++ exercise/kotlin/day13/src/main/kotlin/santamarket/model/ChristmasElf.kt
@@ -2,16 +2,18 @@ package santamarket.model
 
 import santamarket.model.offer.Offer
 import santamarket.model.offer.TenPercentOffer
+import santamarket.model.offer.TwoForAmountOffer
 
 class ChristmasElf(private val catalog: SantamarketCatalog) {
     private val offers = mutableMapOf<Product, Offer>()
 
     fun addSpecialOffer(offerType: SpecialOfferType, product: Product, argument: Double) {
         val offer =
-        if (offerType == SpecialOfferType.TEN_PERCENT_DISCOUNT)
-            TenPercentOffer(offerType, product, argument)
-        else
-            Offer(offerType, product, argument)
+            when (offerType) {
+                SpecialOfferType.TEN_PERCENT_DISCOUNT -> TenPercentOffer(offerType, product, argument)
+                SpecialOfferType.TWO_FOR_AMOUNT -> TwoForAmountOffer(offerType, product, argument)
+                else -> Offer(offerType, product, argument)
+            }
         offers[product] = offer
     }
